Migrate SolrCLI and other sub tools into cli package
