+++ solr/core/src/java/org/apache/solr/cloud/ZkCLI.java
@@ -52,7 +52,7 @@ import org.apache.solr.core.ConfigSetService;
 import org.apache.solr.core.CoreContainer;
 import org.apache.solr.core.NodeConfig;
 import org.apache.solr.core.SolrXmlConfig;
-import org.apache.solr.util.CLIO;
+import org.apache.solr.util.cli.CLIO;
 import org.apache.zookeeper.CreateMode;
 import org.apache.zookeeper.KeeperException;
 import org.xml.sax.SAXException;
