+++ solr/core/src/java/org/apache/solr/packagemanager/PackageManager.java
@@ -59,8 +59,8 @@ import org.apache.solr.packagemanager.SolrPackage.Command;
 import org.apache.solr.packagemanager.SolrPackage.Manifest;
 import org.apache.solr.packagemanager.SolrPackage.Plugin;
 import org.apache.solr.pkg.SolrPackageLoader;
-import org.apache.solr.util.SolrCLI;
 import org.apache.solr.util.SolrVersion;
+import org.apache.solr.util.cli.SolrCLI;
 import org.apache.zookeeper.KeeperException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
