+++ solr/core/src/java/org/apache/solr/util/cli/ConfigSetDownloadTool.java
@@ -26,8 +26,6 @@ import java.util.concurrent.TimeUnit;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
 import org.apache.solr.common.cloud.SolrZkClient;
-import org.apache.solr.util.CLIO;
-import org.apache.solr.util.SolrCLI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
