+++ solr/core/src/java/org/apache/solr/util/cli/CreateTool.java
@@ -25,8 +25,6 @@ import org.apache.commons.cli.Option;
 import org.apache.solr.client.solrj.SolrRequest;
 import org.apache.solr.client.solrj.request.GenericSolrRequest;
 import org.apache.solr.common.util.NamedList;
-import org.apache.solr.util.CLIO;
-import org.apache.solr.util.SolrCLI;
 
 public class CreateTool extends ToolBase {
