+++ solr/core/src/java/org/apache/solr/util/cli/ZkRmTool.java
@@ -25,8 +25,6 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
 import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.common.cloud.SolrZkClient;
-import org.apache.solr.util.CLIO;
-import org.apache.solr.util.SolrCLI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
