+++ solr/core/src/test/org/apache/solr/cloud/PackageManagerCLITest.java
@@ -22,8 +22,8 @@ import java.util.Arrays;
 import org.apache.solr.client.solrj.request.CollectionAdminRequest;
 import org.apache.solr.core.TestSolrConfigHandler;
 import org.apache.solr.util.LogLevel;
-import org.apache.solr.util.PackageTool;
-import org.apache.solr.util.SolrCLI;
+import org.apache.solr.util.cli.PackageTool;
+import org.apache.solr.util.cli.SolrCLI;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.ServerConnector;
