+++ solr/core/src/test/org/apache/solr/cloud/SolrCLIZkUtilsTest.java
@@ -33,9 +33,9 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 import org.apache.solr.common.cloud.SolrZkClient;
 import org.apache.solr.common.cloud.ZkMaintenanceUtils;
-import org.apache.solr.util.SolrCLI;
 import org.apache.solr.util.cli.ConfigSetDownloadTool;
 import org.apache.solr.util.cli.ConfigSetUploadTool;
+import org.apache.solr.util.cli.SolrCLI;
 import org.apache.solr.util.cli.ZkCpTool;
 import org.apache.solr.util.cli.ZkLsTool;
 import org.apache.solr.util.cli.ZkMvTool;
