+++ solr/core/src/test/org/apache/solr/cloud/SolrCloudExampleTest.java
@@ -45,11 +45,11 @@ import org.apache.solr.common.cloud.Slice;
 import org.apache.solr.common.cloud.ZkStateReader;
 import org.apache.solr.common.util.NamedList;
 import org.apache.solr.util.ExternalPaths;
-import org.apache.solr.util.SolrCLI;
 import org.apache.solr.util.cli.ConfigTool;
 import org.apache.solr.util.cli.CreateCollectionTool;
 import org.apache.solr.util.cli.DeleteTool;
 import org.apache.solr.util.cli.HealthcheckTool;
+import org.apache.solr.util.cli.SolrCLI;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
