+++ solr/core/src/test/org/apache/solr/security/BasicAuthIntegrationTest.java
@@ -63,8 +63,8 @@ import org.apache.solr.common.util.TimeSource;
 import org.apache.solr.common.util.Utils;
 import org.apache.solr.embedded.JettySolrRunner;
 import org.apache.solr.util.LogLevel;
-import org.apache.solr.util.SolrCLI;
 import org.apache.solr.util.TimeOut;
+import org.apache.solr.util.cli.SolrCLI;
 import org.apache.solr.util.cli.StatusTool;
 import org.junit.After;
 import org.junit.Before;
