+++ solr/test-framework/src/java/org/apache/solr/cloud/AbstractDistribZkTestBase.java
@@ -39,9 +39,9 @@ import org.apache.solr.common.cloud.ZkStateReader;
 import org.apache.solr.core.Diagnostics;
 import org.apache.solr.core.MockDirectoryFactory;
 import org.apache.solr.embedded.JettySolrRunner;
-import org.apache.solr.util.SolrCLI;
 import org.apache.solr.util.TimeOut;
 import org.apache.solr.util.cli.ConfigSetUploadTool;
+import org.apache.solr.util.cli.SolrCLI;
 import org.apache.zookeeper.KeeperException;
 import org.junit.BeforeClass;
 import org.slf4j.Logger;
